+++ cli/tests/test_git_init.rs
@@ -652,8 +652,8 @@ fn test_git_init_colocated_via_git_repo_path_imported_refs() {
     [EOF]
     ");
 
-    // With remotes.origin.auto-track-bookmarks = ''
-    test_env.add_config("remotes.origin.auto-track-bookmarks = ''");
+    // With remotes.origin.auto-track-bookmarks = '~glob:*'
+    test_env.add_config("remotes.origin.auto-track-bookmarks = '~glob:*'");
     let local_dir = test_env.work_dir("local2");
     set_up_local_repo(local_dir.root());
     let output = local_dir.run_jj(["git", "init", "--git-repo=."]);